chore: move task list components to separate directory
